+++ packages/toast/src/Toast.tsx
@@ -17,54 +17,7 @@ import {
   toastThemeStyles,
   toastTransitionStateStyles,
 } from './styles';
-import { Variant } from './types';
-
-export interface ToastProps extends Omit<React.ComponentProps<'div'>, 'title'> {
-  /**
-   * Optional text shown in bold above the body text.
-   */
-  title?: React.ReactNode;
-
-  /**
-   * Required main text for the Toast.
-   */
-  body: React.ReactNode;
-
-  /**
-   * Optional className passed to the wrapping <div /> for the toast.
-   */
-  className?: string;
-
-  /**
-   * Required style variant to render the Toast as.
-   */
-  variant: Variant;
-
-  /**
-   * Optional number between 0 and 1 that sets the progress bar's progress. Note that the progress bar is only rendered when the Toast variant is set to `'progress'`.
-   *
-   * **Default:** `1`
-   */
-  progress?: number;
-
-  /**
-   * Optional boolean that renders the Toast and makes it visible when true.
-   *
-   * **Default:** `false`
-   */
-  open?: boolean;
-
-  /**
-   * Optional click event handler that, when set, renders a close button that receives the passed handler.
-   */
-  close?: React.MouseEventHandler;
-
-  /**
-   * Determines whether or not the component will be rendered in dark theme.
-   *
-   */
-  darkMode?: boolean;
-}
+import { ToastProps, Variant } from './types';
 
 function Toast({
   title,
@@ -138,9 +91,7 @@ function Toast({
                       data-testid="toast-title"
                       className={cx(
                         currentVariantStyles.body,
-                        css`
-                          font-weight: 700;
-                        `,
+                        baseElementStyles.title,
                       )}
                     >
                       {title}
